+++ src/WebLinterVsix/Commnads/ResetConfigFilesCommand.cs
@@ -17,7 +17,7 @@ namespace WebLinterVsix
             OleMenuCommandService commandService = ServiceProvider.GetService(typeof(IMenuCommandService)) as OleMenuCommandService;
 
             var menuCommandID = new CommandID(PackageGuids.ConfigFileCmdSet, PackageIds.ResetConfigFiles);
-            var menuItem = new OleMenuCommand(ResetConfigurationFiles, menuCommandID);
+            var menuItem = new OleMenuCommand(async (s, e) => { await ResetConfigurationFiles(s, e); }, menuCommandID);
             commandService.AddCommand(menuItem);
         }
 
@@ -33,14 +33,14 @@ namespace WebLinterVsix
             Instance = new ResetConfigFilesCommand(package);
         }
 
-        private void ResetConfigurationFiles(object sender, EventArgs e)
+        private async System.Threading.Tasks.Task ResetConfigurationFiles(object sender, EventArgs e)
         {
             string msg = "This will reset the configuration for all the linters to their defaults.\n\nDo you wish to continue?";
             var result = MessageBox.Show(msg, Constants.VSIX_NAME, MessageBoxButtons.YesNo, MessageBoxIcon.Question);
 
             if (result == DialogResult.Yes)
             {
-                LinterService.EnsureDefaults(true);
+                await LinterService.EnsureDefaults(true);
                 WebLinterPackage.Settings.ResetSettings();
                 WebLinterPackage.Dte.StatusBar.Text = "Web Linter configuration files have been reset";
                 Telemetry.TrackEvent($"VS Reset Configs");